Splitting the Unit tests for TComponent into separate files based on what is tested
